chore: major refactor and cleanup in backend apis
